Fix bug where log term appended wrong
With the current implementation, the follower appends in the log entry the current term of the leader. In the case where the leader appends some logs and a node is down, then when the other node wakes if the leader has changed it would append in the logs the term of the new leader and the logs would not be consistent. Now the append entries RPC includes the whole entry instead of only the commands so the follower appends the correct term of the old leader (the one that sent the initial request)