Fixed raft node never changing voted_for
Because the raft nodes never changed voted for this caused multiple issues. Most likely one of them is #9. If they all started at almost exactly the same time and voted for themselves they would never change back and this will never end. This probably why it happed so rarely.
close #1 for now and we will see.